Create seperate client & server handles for the channel
