+++ experimental/oak_baremetal_runtime/src/framing.rs
@@ -18,7 +18,9 @@ use crate::remote_attestation::AttestationHandler;
 use alloc::vec::Vec;
 use anyhow::Context;
 use ciborium_io::{Read, Write};
-use oak_baremetal_communication_channel::{schema, schema::TrustedRuntime, InvocationChannel};
+use oak_baremetal_communication_channel::{
+    schema, schema::TrustedRuntime, server::ServerChannelHandle,
+};
 use oak_idl::Handler;
 use oak_remote_attestation::handshaker::{
     AttestationBehavior, AttestationGenerator, AttestationVerifier,
@@ -90,12 +92,12 @@ where
         attestation_handler,
     }
     .serve();
-    let invocation_channel = &mut InvocationChannel::new(channel);
+    let channel_handle = &mut ServerChannelHandle::new(channel);
     loop {
-        let message = invocation_channel
-            .read_message()
+        let message = channel_handle
+            .read_request()
             .context("couldn't receive message")?;
         let response = invocation_handler.invoke((&message).into());
-        invocation_channel.write_message(response.into())?
+        channel_handle.write_response(response.into())?
     }
 }
